Add workaround for csv Sniffer issue
The `Sniffer` in the `csv` Python module analyses the frequency
of characters to infer the CSV file's dialect.  This often goes
wrong when the CSV file has but a single column.
In some circumstances the Sniffer raises an exception, but not
always.  Workaround is to limit the set of possible delimiter
characters.