Fix PyInstaller script location issue
Originally, worker-ng used the location of the Python script
to determine the installation directory for reading configuration
files and templates.
When using PyInstaller, this no longer works as the path of
the script is a temporary directory.
The installation directory as set when building with CMake
is now hardcoded into the relevant Python scripts.